fix(v4): use long/DiscordUser for slash options (int/DiscordMember unsupported)
DSharpPlus 4 SlashCommands matches option types by exact equality: int and
DiscordMember aren't allowed, which crashed registration for userlimit, limit,
bitrate, volume, notify remove, and voice permit/kick/transfer. Switch int
options to long (cast back where needed) and DiscordMember options to
DiscordUser (resolve the member via GetMemberAsync).

+++ src/DiscordBot/Discord/Commands/ConfigCommands.cs
@@ -54,10 +54,10 @@ public sealed class ConfigCommands : ApplicationCommandModule
     [SlashCommand("userlimit", "Задать предел гостей для новых временных залов (0 = без предела).")]
     public async Task UserLimitAsync(
         InteractionContext ctx,
-        [Option("limit", "Предел гостей по умолчанию, 0-99.")] int limit)
+        [Option("limit", "Предел гостей по умолчанию, 0-99.")] long limit)
     {
         limit = Math.Clamp(limit, 0, 99);
-        await UpdateAsync(ctx.Guild!.Id, c => c.DefaultUserLimit = limit);
+        await UpdateAsync(ctx.Guild!.Id, c => c.DefaultUserLimit = (int)limit);
         await ctx.ReplyAsync($"Ура! Отныне каждый чертог примет **{(limit == 0 ? "без счёта" : limit.ToString())}** доблестных товарищей!", ephemeral: true);
     }
